fileSD: Do not try to remove non-existent lease file
The volume.lease file exists only if a domain has volume leases. Avoid
removing the non-existent lease file and the bogus exception in the log
when the operation "fails".

This is the same patch from master, applied to FileStorageDomain instead
of FileStorageDomainManifest, since deleteImage was not moved to the
manifest class in 3.6.

Bug-Url: https://bugzilla.redhat.com/1291207
Reviewed-on: https://gerrit.ovirt.org/50589
Continuous-Integration: Jenkins CI
Reviewed-on: https://gerrit.ovirt.org/51132

+++ vdsm/storage/fileSD.py
@@ -401,9 +401,10 @@ class FileStorageDomain(sd.StorageDomain):
                 metaFile = volPath + '.meta'
                 self.log.debug("Removing file: %s", metaFile)
                 self.oop.os.remove(metaFile)
-                leaseFile = volPath + '.lease'
-                self.log.debug("Removing file: %s", leaseFile)
-                self.oop.os.remove(leaseFile)
+                if self.hasVolumeLeases():
+                    leaseFile = volPath + '.lease'
+                    self.log.debug("Removing file: %s", leaseFile)
+                    self.oop.os.remove(leaseFile)
             except OSError:
                 self.log.error("vol: %s can't be removed.",
                                volPath, exc_info=True)
